I have moved `KeysFragment` and `SymmetricKeyRecyclerViewAdapter` and their layouts to the `app` module.
