+++ app/src/main/res/layout/sym_listitem.xml
